Converting double to BigDecimal

+++ money-transfer-domain/src/main/java/com/lmartino/bank/domain/usecase/CreateAccountUseCase.java
@@ -4,6 +4,8 @@ import com.lmartino.bank.domain.adapter.AccountRepository;
 import com.lmartino.bank.domain.model.Account;
 import com.lmartino.bank.domain.model.Amount;
 
+import java.math.BigDecimal;
+
 public class CreateAccountUseCase {
     private AccountRepository accountRepository;
 
@@ -11,7 +13,7 @@ public class CreateAccountUseCase {
         this.accountRepository = accountRepository;
     }
 
-    public Account compose(final String name, final double balance) {
+    public Account compose(final String name, final BigDecimal balance) {
         Account account = Account.createNewAccount(name, Amount.of(balance));
         return accountRepository.saveAccount(account);
     }
